Update README.md
"I'm an aspiring Data Analyst with a passion for turning numbers into stories. I thrive on solving puzzles hidden within data and transforming them into valuable insights. With a strong foundation in mathematics and a hunger for knowledge, I'm on a mission to make data work its magic for informed decisions and business success."